Add ability to set a stripe
The new plan is to have a small set of members (<=7).  To decide who is the leader, we'll set a stripe across the top of the display of length equal to the leader's index.  Each member will present their color as a pixel, and we'll be able to tell which of those members is acting as leader.