Initial rust implementation
This implementation functions over the real network.  It allows for clients to send in commands, and for the commands to be replicated and applied.  In the in-memory version (see main.rs) there is a mechanism for testing isolated clients.  In the real-world version (see launch.sh), the firewall commands do not yet exist.

This peforms well on my mac, but doesn't work on the raspberry pi.  It generally functions, but the networking appears to be terrible, and messages are being delayed like crazy.  In one run, the leader was seeing ACKs for records that the follower sent 20 seconds prior.  Consequently the leader is retrasmitting a whole bunch.  The entire thing could benefit from connection pooling I suspect, but that shouldn't be required for this low message rate.  Slowing things down didn't help: client connections timeout and re-election happens often.